chore(secret): fold variable-name tests into secret_test.go
